+++ c/src/connections/client.h
@@ -93,6 +93,6 @@ enum query_error {
 extern const rapidjson::GenericStringRef<char> query_error_text[];
 
 void* client_connection_handle(void* arg);
-int send_ka(client_socket_data* socket_data);
+ssize_t send_ka(client_socket_data* socket_data);
 void send_res(client_socket_data* socket_data, const char* data, uint32_t length);
 void send_json(client_socket_data* socket_data, rapidjson::Document& data);
